予定・実績時間の方をfloatに変更

+++ models.go
@@ -170,8 +170,8 @@ type Issue struct {
 	Milestone      []*Version     `json:"milestone,omitempty"`
 	StartDate      time.Time      `json:"startDate,omitempty"`
 	DueDate        time.Time      `json:"dueDate,omitempty"`
-	EstimatedHours int            `json:"estimatedHours,omitempty"`
-	ActualHours    int            `json:"actualHours,omitempty"`
+	EstimatedHours float32        `json:"estimatedHours,omitempty"`
+	ActualHours    float32        `json:"actualHours,omitempty"`
 	ParentIssueID  int            `json:"parentIssueId,omitempty"`
 	CreatedUser    *User          `json:"createdUser,omitempty"`
 	Created        time.Time      `json:"created,omitempty"`
